+++ ios/PlanetArchitect/PlanetArchitectApp.swift
@@ -1,0 +1,10 @@
+import SwiftUI
+
+@main
+struct PlanetArchitectApp: App {
+    var body: some Scene {
+        WindowGroup {
+            LevelSelectView()
+        }
+    }
+}